fix:파일구조 수정
