--- database-application/README.md
@@ -1,1 +1,0 @@
-# database-application
