--- database-application/src/main/java/com/example/database/DatabaseApplication.java
@@ -1,13 +1,0 @@
-package com.example.database;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DatabaseApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DatabaseApplication.class, args);
-    }
-
-}
