--- database-application/src/main/java/com/example/database/NewGenerationController.java
@@ -1,14 +1,0 @@
-package com.example.database;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class NewGenerationController {
-    @GetMapping(path="/")
-    public String getPerson() {
-        return "hello1";
-
-    }
-}
