--- database-application/src/main/java/com/example/database/controller/WelcomeController.java
@@ -1,12 +1,0 @@
-package com.example.database.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class WelcomeController {
-    @GetMapping("/main")
-    public String welcome(){
-        return "hello1";
-    }
-}
