--- database-application/src/main/java/com/example/database/dto/NaverRequestVariableDto.java
@@ -1,8 +1,0 @@
-package com.example.database.dto;
-
-import lombok.Data;
-
-@Data
-public class NaverRequestVariableDto {
-    private String query;
-}
