--- database-application/src/main/java/com/example/database/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.database.repository;
-
-import com.example.database.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
