--- database-application/src/main/resources/application.properties
@@ -1,34 +1,0 @@
-spring.application.name=database
-
-spring.jpa.generate-ddl=true
-
-spring.jpa.hibernate.ddl-auto=update
-
-
-spring.datasource.url=jdbc:oracle:thin:@114.70.92.70:1521:xe
-
-spring.datasource.username=book_203604
-
-spring.datasource.password=book_203604
-
-#spring.datasource.url=jdbc:oracle:thin:@localhost:1521:xe
-
-#spring.datasource.username=c##dba
-
-#spring.datasource.password=1234
-
-spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
-
-spring.thymeleaf.prefix=classpath:/templates/
-
-server.port=8082
-
-naver.client-id=7BuQwVjiMWaKTldjEG5W
-naver.client-secret=k99twpIy8y
-
-ebay.api.access_token=v^1.1#i^1#r^0#p^3#f^0#I^3#t^H4sIAAAAAAAAAOVZf2wbVx2P8wtFWzvKuqUMhFy3XVmrs9/9sO/HGhc3dmq3cX7YTliDWPTu7p3zmvOde+9dEnedSAKaGNofCE0IBGgRo1KH9kfZ/qhE1VGBEAgJhR8FgSZYN421FJCQ6DQqisSdk7huWNskjpgl/I91776/Pt+f994Ds51d+55JP/PelsCHWhdmwWxrIMDeB7o6O/ZvbWt9pKMF1BEEFmZ3z7bPt109QGDJLCs5RMq2RVBwpmRaRKku9oRcx1JsSDBRLFhCRKGakk9k+xUuDJSyY1Nbs81QMJPsCXFRWeRiUV7So6ygizFv1VqRWbB7QhKUDVbWDBRFSJV01XtPiIsyFqHQoh4/4ASGZRlOKABe4WQFcOEoEMdCwVHkEGxbHkkYhOJVc5Uqr1Nn691NhYQgh3pCQvFMoi8/mMgkUwOFA5E6WfFlP+QppC65/anX1lFwFJouursaUqVW8q6mIUJCkfiShtuFKokVYzZgftXVKgdYA6qargpI5XVuU1zZZzslSO9uh7+CdcaokirIophW7uVRzxvqcaTR5acBT0QmGfT/hl1oYgMjpyeUOpQ4NpJP5ULB/NCQY09hHenVpIpJksSKHOdZaxYn5HFOXFaxJGfZwat09NqWjn13keCATQ8hz1602itsnVc8okFr0EkY1Lelni667D1Blsb8cC7Fz6UTlh9RVPJcEKw+3tv3K8lwK/yblQ6GCDjEQZGTBEmNivod0sGv9XWlRNyPSmJoKOLbglRYYUrQmUS0bEINMZrnXreEHKwrfNTgeMlAjB6TDUaQDYNRo3qMYQ2EgFftqiZL/x+ZQamDVZeiWnasflGF1xPKa3YZDdkm1iqh1STVPrOcCzOkJzRBaVmJRKanp8PTfNh2ihEOADbyRLY/r02gEgzVaPG9iRlczQoNeVwEK7RS9qyZ8ZLOU24VQ3He0YegQyt5ZJrewkrK3mZbfPXqHUD2mtjzQMFT0VwY0zahSG8Imo6msIbGsd5cyDiOkwVekIBf6zIAckMgTbuIrSyiE3aTwTw8OHi4P9UQNq99QtpcqGrdhS9wYLkLCTGJAaICQENgE+VyplRyKVRNlGmyWHp9VuTYhuCVXbfZCtEtGcCunJRPNNhC/amrYGgo1J5E1vu3Ur/WP0isuVRfLpVPjxcGj6YGGkKbQ4aDyETBx9pseZoYThxJeL9stth/LGZWJvsM2UxnMqLOugSQisrCUgEVnEJlhBYi4hOl9MxMMomFYjmHQLRyOJInR04gLoqLPT0NOSmPNAc1WeuS9aiWE9Ol/fnoob5oX6bfMfjodGZIGhxV5f2jvcPG8TRKTA3nxjKNgc8Wm63SayO34XFbuHOJ1wD6tf4BgHSWCnO82oXGvaeGgKaKTdevoxrLGrLAsjIPoCpAUTZElZeRYXj/Rqyxj0V//DYZ3qPHRgYOp1ODDCQ6M5RLMoaBWF6VobeHkzhD0hHf4Ehutghv1kQm/s7tfwLNr/U1w/NlEE8ILOOw/9EQ1uxSxIYunfCXxqtWB9dCFCHezi+8tNH3JIcdBHXbMisbYV4HD7amvL2i7VQ2orDGvA4eqGm2a9GNqFtmXQeH4ZoGNk3/QGAjCuvY12OmBc0KxRrZkEps+dlG1sFShpUqQB2Tsl8va+L01krI0VAY60vHiRsx1kGeQlg9QtsI0zpV1ky2bIoNrC3JIK5KNAeX125FvRy/1t9f1kb8QbxaWFfolhhqqhrbWSMdO0ij466Dm2sErAy9gfH+VIpZNQIZWjmpQaihkw2h973bjEcmmeQm7M+SaKrZPmQEEQFZlwVGk6M6I0BOZqAkSIwosZwgSFEOxsSGMG/+MVH73MsNgmZFnhNYSebWfCy0aqHuePq/7iQit18HxluqP3Y+8EMwH3itNRAAB8AedhfY2dk20t52/yMEU697QyNMcNGC1HVQeBJVyhA7rQ+2LG7t1+fS/e/Oqu65T18/KLVsqbuNXPgs2FG7j+xqY++ru5wEH7/1poN9oHuLB9iLKOA5GXBjYNett+3sw+3bv7HvoevfiW+bFPV/3N/d/fLpBfnm62BLjSgQ6Ghpnw+0PDqXf/q357c/3vHCG22vHrn41yevXOns+pV848qNyIsPX/pm8Mt//3f8c88+9/sfff7bhX/uevv0u4g7HtrxqTeD45N7n3vn6yeffr471vnSJ36ZLCT3PX72UH7bH3/x65987NRH3MrO/vMdu78KL2L+1A/0vV+8PPJW+mtn9wSEgw+eO3v1qasFuuuaIc29/ubYtZf+tFdb5H4Md/x824tPfutfP5u7Ofm9Mzf4xzK9C9fOX39t4ZP03Pd35hY/Ghj57kMftvJvXH7+zOXPpHPbne6/fCF74iu/e+o3p+grNy8c3XPhhS9deuW90ZHFgxcf+0PXn/mfXjiy2N2jTb2d7NudP7P70b8Jleyrw5fmt751Wn4ne4HLPGAsxfI/vntFtCceAAA=ebay.api.refresh_token=v^1.1#i^1#r^1#p^3#f^0#I^3#t^Ul4xMF8xMToyRkFENzNFNjg4QkZGREFCNjAzMzNFRDhCMTM1RjQxRl8wXzEjRV4yNjA=
-ebay.api.refresh_token=v^1.1#i^1#r^1#p^3#f^0#I^3#t^Ul4xMF8xMToyRkFENzNFNjg4QkZGREFCNjAzMzNFRDhCMTM1RjQxRl8wXzEjRV4yNjA=
-
-ebay.client.id=KYUNGHEO-asd-PRD-ffe13b9ae-d82f8de3
-ebay.client.secret=PRD-fe13b9ae1e93-eea4-4703-9d58-db21
-ebay.api.url=https://api.ebay.com
