--- database-application/src/main/resources/templates/hello1.html
@@ -1,6 +1,0 @@
-<html xmlns:th="http://www.thymeleaf.org">
-<body>
-<h2>로그인 성공</h2>
-<p>환영합니다! 로그인에 성공하셨습니다.</p>
-</body>
-</html>
