--- database-application/src/main/resources/templates/login.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Login</title>
-</head>
-<body>
-<h2>Login</h2>
-<div th:if="${param.error}">
-    <p style="color:red;">로그인에 실패했습니다. 이메일 또는 비밀번호를 확인해주세요.</p>
-</div>
-<form th:action="@{/users/login}" method="post">
-
-    <label for="email">Email:</label>
-    <input type="email" id="email" name="email" required><br>
-
-    <label for="password">Password:</label>
-    <input type="password" id="password" name="password" required><br>
-
-    <button type="submit">Login</button>
-
-</form>
-<a href="/users/register"><button>회원가입</button></a>
-</body>
-</html>
