--- database-application/src/main/resources/templates/register.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Register</title>
-</head>
-<body>
-<h2>Register</h2>
-<form th:action="@{/users/register}" method="post">
-    <label for="email">Email:</label>
-    <input type="email" id="email" name="email" required><br>
-
-    <label for="username">Username:</label>
-    <input type="text" id="username" name="username" required><br>
-
-    <label for="password">Password:</label>
-    <input type="password" id="password" name="password" required><br>
-
-    <button type="submit">Register</button>
-</form>
-</body>
-</html>
