--- src/main/resources/application.properties
@@ -1,31 +1,0 @@
-spring.application.name=database
-
-spring.jpa.generate-ddl=true
-
-spring.jpa.hibernate.ddl-auto=update
-
-
-spring.datasource.url=jdbc:oracle:thin:@studyswh.synology.me:49158:xe
-
-spring.datasource.username=kang
-
-spring.datasource.password=123
-
-#spring.datasource.url=jdbc:oracle:thin:@localhost:1521:orcl
-
-#spring.datasource.username=c##lepied
-
-#spring.datasource.password=1234
-
-spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
-
-spring.thymeleaf.prefix=classpath:/templates/
-
-server.port=8082
-
-naver.client-id=7BuQwVjiMWaKTldjEG5W
-naver.client-secret=k99twpIy8y
-
-spring.jpa.hibernate.use-new-id-generator-mappings=true
-spring.jpa.properties.hibernate.id.new_generator_mappings=true
-
